requirements.txt
Flask==3.0.0
argon2-cffi==23.1.0
pyotp==2.9.0
markdown==3.5.1
bleach==6.1.0
qrcode==7.4.2
gunicorn==21.2.0
pycryptodome==3.20.0
